# Improve performance of presenting ChatChannelVC