### 🔗 Issue Links

Resolves: [PBE-6192](https://stream-io.atlassian.net/browse/PBE-6192)

### 🎯 Goal

Present `ChatChannelVC` quicker

### 📝 Summary

- Reduce cell reloads after applying message list changes
- Skip duplicate `updateContent` call when creating the `ChatMessageContentCell`

### 🛠 Implementation

`ChatMessageContentView` parses markdown and links which can get expensive. Reloading a cell repeats the parsing. This PR aims to reduce cell reloads. 

### 🎨 Showcase

Measured with iPhone 12 Pro Max and 3 different channels using a separate sample app what just presents the `ChatChannelVC` in a sheet. Time spent by the main thread.

Channel with 25 simple messages with text. Negligible change here. **5-10%** faster.
| Criteria | Before | After |
| ----- | ------ | ----- |
| Main thread |  269 ms  |  251 ms  |
| Hang time | 297 ms  | 291 ms  |

Channel with complex markdown and some other simple messages with links. **~50%** faster
| Criteria | Before | After |
| ----- | ------ | ----- |
| Main thread |  969 ms  |  420 ms  |
| Hang time | 811 ms  | 480 ms |

All of this really depends on how much processing the message cell requires. If more, the impact is higher because it is less time recalculated.

### 🧪 Manual Testing Notes

##### Verify sending giphys
1. Open giphy preview
2. Send/cancel a giphy

##### Verify deleting messages and displaying user avatars.

1. Log in from 2 devices/simulators
2. Open the same channel with 2 different users
3. Send 2 messages: "A" and "B"
4. Hard delete the "B"
Result: Avatar is shown for message "A"

##### Verify sending messages and displaying user avatars.
1. Send message "A" > avatar is shown for it
2. Send message "B" > avatar is shown for "B" and not for "A" since messages are grouped

### ☑️ Contributor Checklist

- [x] I have signed the [Stream CLA](https://docs.google.com/forms/d/e/1FAIpQLScFKsKkAJI7mhCr7K9rEIOpqIDThrWxuvxnwUq2XkHyG154vQ/viewform) (required)
- [x] This change should be manually QAed
- [x] Changelog is updated with client-facing changes
- [ ] Changelog is updated with new localization keys
- [x] New code is covered by unit tests
- [ ] Comparison screenshots added for visual changes
- [ ] Affected documentation updated (docusaurus, tutorial, CMS)


[PBE-6192]: https://stream-io.atlassian.net/browse/PBE-6192?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ